finished the user crud? finally?

+++ src/main/java/org/slothmq/server/web/dto/Paged.java
@@ -3,8 +3,8 @@ package org.slothmq.server.web.dto;
 import java.util.List;
 
 public record Paged<T>(List<T> content,
-                       int totalPages,
-                       int totalElements) {
+                       long totalPages,
+                       long totalElements) {
 
     public boolean isEmpty() {
         return content.isEmpty();
